continue service.finish UserService and PowerBankService

+++ Portable Power Bank Rental System/src/main/java/powerbankrental/dao/UserDAO.java
@@ -4,12 +4,11 @@ import powerbankrental.model.User;
 import powerbankrental.model.enums.Role;
 
 import java.sql.Connection;
-import java.util.Date;
 import java.util.List;
 
 //用户数据访问接口-定义用户相关的数据库操作方法
 public interface UserDAO {
-    int addUser(User user);    //添加用户
+    long addUser(User user);    //添加用户
     int addUserInTransaction(User user, Connection connection);
 
     User getUserById(long userId);    //根据用户ID查询用户
